Enable RBE testing of JAX CUDA 13 wheels by loading CUDA UMD libs from Python code.
Testing of CUDA 13 wheels was not possible on RBE machines because the tests used CUDA 12 driver installed on RBE VMs. To fix that, the tests that are executed with `build_jaxlib=false` or `build_jaxlib=wheel` will use hermetic CUDA UMD libs.

If `--@cuda_driver//:include_cuda_umd_libs=true`:

1) Zip CUDA UMD files and add the archive in `zip_deps` attribute of `jax_cuda_plugin_py_import` and `pypi_jax_cuda_plugin_with_cuda_deps`.

2) In Bazel environment, load UMD libs in `jax_plugins/cuda/__init__.py`.
  This approach force-loads the libraries from hermetic UMD before any other lib loads it dynamically.

PiperOrigin-RevId: 831570901